+++ elements/001-hydrogen/hydrogen/src/utils/utils_queue.h
@@ -13,7 +13,8 @@
 
 #include <stddef.h>
 
-#include "../config/configuration.h"
+#include "../config/config_forward.h"  // For AppConfig forward declaration
+#include "../config/config_structs.h"  // For queue-related constants
 
 // Default values for early initialization
 #define MAX_QUEUE_BLOCKS 128  // Maximum possible blocks