+++ elements/001-hydrogen/hydrogen/src/utils/utils_status.h
@@ -13,7 +13,7 @@
 
 #include <jansson.h>
 #include <time.h>
-#include "../config/configuration.h"
+#include "../config/config.h"
 
 // File descriptor information structure
 typedef struct {