+++ elements/001-hydrogen/hydrogen/src/websocket/websocket_server_status.c
@@ -58,7 +58,7 @@
 #include "websocket_server.h"
 #include "websocket_server_internal.h"
 #include "../logging/logging.h"
-#include "../config/configuration.h"
+#include "../config/config.h"
 #include "../utils/utils.h"
 
 extern AppConfig *app_config;